Model for delivery/backend/requirements.txt
# TAFE Leak Detection Backend - Requirements
# FastAPI and ASGI
fastapi>=0.109.0
uvicorn[standard]>=0.25.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0

# Settings
pydantic>=2.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0

# Utilities
python-dotenv>=1.0.0
